cpp/2019/intcode: parse integers using `stol` instead of `stoi`
Since we are now using 64-bit integers, we need to parse them as `long
int`s as well.

+++ cpp/year2019/intcode/intcode.cc
@@ -219,7 +219,7 @@ namespace intcode {
     memory m;
     m.reserve(parts.size());
     for (auto part : parts) {
-      m.push_back(std::stoi(part));
+      m.push_back(std::stol(part));
     }
     return m;
   }
